[expo-router] Fix Stack.Screen options not accepting webModalStyle in TypeScript
The ExtendedStackNavigationOptions type that includes webModalStyle was
defined in StackClient.tsx but StackScreen.tsx used NativeStackNavigationOptions
directly for its StackBaseScreenProps, causing the public Stack.Screen options
prop to not accept webModalStyle.

Move ExtendedStackNavigationOptions (and the new WebModalStyle interface) to
StackScreen.tsx so they are part of the StackScreenProps.options type.
Re-export ExtendedStackNavigationOptions from StackClient.tsx to keep
existing internal importers working.

Fixes #44485
